feat: subir alteracoes

+++ src/main/java/org/example/models/PacienteModel.java
@@ -5,10 +5,9 @@ import jakarta.persistence.*;
 @Entity
 @Table(name = "Paciente")
 public class PacienteModel {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     private String nome;
     private String referencia;
@@ -23,7 +22,7 @@ public class PacienteModel {
 
     public long getId() { return id; }
 
-    public void setId(long id) { this.id = id; }
+    public void setId(int id) { this.id = id; }
 
     public String getNome() {
         return nome;
